+++ yash-env/src/lib.rs
@@ -530,7 +530,7 @@ impl Env {
     }
 }
 
-mod alias;
+pub mod alias;
 pub mod any;
 pub mod builtin;
 mod decl_util;